variety of fixes, mainly refactoring models
* refactor models from __init__ to separate modules
* base64 encode payload hash
* remove workflow_version from payload responses
* improvements to process model generation methods and properties
* remove metric models
* cleanup/fixes for typing errors
* make some model params required
* response_model_exclude_unset=True to filter unset model fields
* random formatting fixes
