+++ tests/api/test_workflows.py
@@ -20,8 +20,8 @@ def test_loading_workflows(settings):
     name = "mirror"
     workflows = Workflows.from_yaml(settings.workflow_config_file)
     assert name in workflows
-    assert hasattr(workflows[name], "name")
-    assert workflows[name].name == name
+    assert hasattr(workflows[name], "id")
+    assert workflows[name].id == name
 
 
 def test_loading_workflows_bad_config_file(bad_workflow_config):
