+++ workflow-config.yml
@@ -4,18 +4,18 @@ workflows:
     version: 2
     handler: argo-handler
     type: argo-workflow
-    argo_template: workflowtemplate/mirror-workflow
-    cache_key_hash_includes:
+    argoTemplate: workflowtemplate/mirror-workflow
+    cacheKeyHashIncludes:
       - .features[].id
       - .features[].collection
-    cache_key_hash_excludes: []
+    cacheKeyHashExcludes: []
   cirrus-example:
     description: "An example workflow config for a cirrus workflow"
     version: 1
     handler: cirrus-handler
     type: cirrus-workflow
-    sfn_arn: arn:aws:states:us-west-2:09876543210:stateMachine:cirrus-example
-    cache_key_hash_includes:
+    sfnArn: arn:aws:states:us-west-2:09876543210:stateMachine:cirrus-example
+    cacheKeyHashIncludes:
       - .features[].id
       - .features[].collection
-    cache_key_hash_excludes: []
+    cacheKeyHashExcludes: []
